fuck allwinner

+++ build.sh
@@ -54,6 +54,7 @@ cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
   ..
 
 if [ "$BUILD" = "ON" ]; then
+  echo " >> Building:"
   cmake --build . -- -j$(nproc)
 fi
 
@@ -62,20 +63,27 @@ if [ -f compile_commands.json ]; then
 fi
 
 if [ "$TESTS" = "ON" ]; then
+  echo " >> Running tests:"
   ctest -V --output-on-failure
 fi
 
 if [ "$NEW_SD" = "ON" ]; then
-  rm sdcard.img
+  echo " >> Creating new SD card image:"
+  if [ -f sdcard.img ]; then
+    rm sdcard.img
+  fi
   dd if=/dev/zero of=sdcard.img bs=1M count=64
   dd if=../u-boot-sunxi-with-spl.bin of=sdcard.img bs=1024 seek=8 conv=notrunc
 fi
 
 if [ "$UPLOAD" = "ON" ]; then
+  echo " >> Uploading to SD card:"
+  #dd if=bootloader-egon.bin of=sdcard.img bs=1024 seek=40 conv=notrunc
   dd if=bootloader.bin of=sdcard.img bs=1024 seek=40 conv=notrunc
   dd if=kernel.bin of=sdcard.img bs=1024 seek=1024 conv=notrunc
 fi
 
 if [ "$QEMU" = "ON" ]; then
-  qemu-system-arm -M orangepi-pc -drive file=sdcard.img,format=raw,if=sd -serial stdio
+  echo " >> Starting QEMU:"
+  qemu-system-arm -M orangepi-pc -drive file=sdcard.img,format=raw,if=sd -monitor telnet:127.0.0.1:4444,server,nowait
 fi
